+++ marketplace/src/main/java/com/boost/marketplace/ui/details/FeatureDetailsActivity.kt
@@ -1932,7 +1932,7 @@ class FeatureDetailsActivity :
             }).startAnimation()
     }
 
-    private fun getDiscountedPrice(price: Double, discountPercent: Int): Double {
+    private fun getDiscountedPrice(price: Double, discountPercent: Double): Double {
         return price - ((discountPercent / 100) * price)
     }
